Initial implementation of value types bugfix
This is a first attempt at completing the implementation of value
types. This implementation is very inline with a suggestion from
a contributor (@victorandree) that can be found here:
https://github.com/apollographql/apollo-server/issues/3141#issuecomment-521677383

The solution is relatively straightforward, though a bit hacky.
After the maps are constructed, we go back and "clean them up"
by identifying value types in the map and removing any signs of
service ownership from them.

While building a query plan, the gateway now sees an "unowned type"
as belonging to the parentGroup's service. I'm unsure if this has
any reaching implications (previously an unowned type or field would
throw), however this code path was never exercised in our test cases.